% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_network.R
\name{analyze_network,omics_network-method}
\alias{analyze_network,omics_network-method}
\alias{analyze_network}
\alias{analyze_network-methods}
\title{Analysing the network}
\usage{
\S4method{analyze_network}{omics_network}(Omega, nv, label_v = NULL)
}
\arguments{
\item{Omega}{a omics_network object}

\item{nv}{the level of cutoff at which the analysis should be done}

\item{label_v}{(optionnal) the name of the genes}
}
\value{
A matrix containing, for each node, its betweenness,its degree, its
output, its closeness.
}
\description{
Calculates some indicators for each node in the network.
}
\examples{

data(network)
analyze_network(network,nv=0)

}
\author{
Bertrand Frederic, Myriam Maumy-Bertrand.
}
\keyword{methods}
