\name{dim}
\docType{methods}
\alias{dim}
\alias{dim-methods}
\alias{dim,micro_array-method}
\title{Dimension of the data}
\description{
Dimension of the data
}
\section{Methods}{
\describe{

\item{\code{signature(x = "micro_array")}}{
Gives the dimension of the matrix of measurements.
}
}}
\keyword{methods}

