% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro_array.R
\docType{methods}
\name{dim}
\alias{dim}
\alias{dim-methods}
\alias{dim,micro_array-method}
\title{Dimension of the data}
\usage{
\S4method{dim}{micro_array}(x)
}
\arguments{
\item{x}{an object of class `micro_array`.}
}
\description{
Dimension of the data
}
\section{Methods}{
 \describe{

\item{list("signature(x = \"micro_array\")")}{ Gives the dimension of the
matrix of measurements. } }
}

\author{
Bertrand Frederic, Myriam Maumy-Bertrand.
}
\keyword{methods}
