% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{doc}
\alias{doc}
\title{Human transcription factors from HumanTFDB}
\description{
Retrieve human transcription factors from HumanTFDB, extracted from 
AnimalTFDB 3.0: a comprehensive resource for annotation and prediction 
of animal transcription factors. Hui Hu, Ya-Ru Miao, Long-Hao Jia, 
Qing-Yang Yu, Qiong Zhang and An-Yuan Guo. *Nucl. Acids Res*. (2018).
}
\examples{

data(doc)
head(doc)
summary(doc)

}
\author{
Bertrand Frederic, Myriam Maumy-Bertrand.
}
\keyword{datasets}
