\name{geneNeighborhood}
\alias{geneNeighborhood}
\alias{geneNeighborhood-methods}
\alias{geneNeighborhood,network-method}
\title{Find the neighborhood of a set of nodes.}

\description{
Find the neighborhood of a set of nodes.
}
\usage{
	geneNeighborhood(net,targets,...)
}

\arguments{
 \item{net}{a network object}
 \item{targets}{a vector containing the set of nodes}
 \item{...}{Optional arguments. See plot options. }
}

\value{
  The neighborhood of the targeted genes. 
}

\author{
Bertrand Frederic, Myriam Maumy-Bertrand.
}



\examples{
	#See vignette
}

\keyword{methods}
