% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro_array.R
\name{head,micro_array-method}
\alias{head,micro_array-method}
\alias{methods}
\alias{head-methods}
\alias{head,ANY-method}
\title{Overview of a micro_array object}
\usage{
\S4method{head}{micro_array}(x, ...)
}
\arguments{
\item{x}{an object of class `micro_array`.}

\item{...}{additional parameters}
}
\description{
Overview of a micro_array object.
}
\section{Methods}{
 \describe{

\item{list("signature(x = \"ANY\")")}{ Gives an overview. }

\item{list("signature(x = \"micro_array\")")}{ Gives an overview. } }
}

\examples{

  if(require(CascadeData)){
	data(micro_US)
	micro_US<-as.micro_array(micro_US,time=c(60,90,210,390),subject=6)
	head(micro_US)
	}

}
\author{
Bertrand Frederic, Myriam Maumy-Bertrand.
}
\keyword{methods}
