\name{infos}
\alias{infos}
\docType{data}
\title{
Details on some probesets of the affy_hg_u133_plus_2 platform.
}
\description{
Dataset with information on the affy_hg_u133_plus_2 platform such as probeset name (affy_hg_u133_plus_2), ensembl_gene_id, entrezgene, hgnc_symbol, chromosome_name, start_position, end_position and band.
}
\usage{data("infos")}
\format{
  The format is:
 chr "infos"
}
\details{
Data.frame with 8859 rows and 8 variables.
}
\author{
Bertrand Frederic, Myriam Maumy-Bertrand.
}

\examples{
data(infos)
}
\keyword{datasets}
