% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{jetsetscores}
\alias{jetsetscores}
\title{jetsetscoresHuman transcription factors from HumanTFDB}
\description{
This data set provides gene target and quality scores for each probe set 
on the hgu133plus2 Affymetrix gene expression microarrays.
}
\details{
Scores are calculated from BLASTN alignments between probe sequences and 
Refseq transcript sequences, as described in the vignette and in the 
reference below.

The Refseq human RNA was downloaded from NCBI on 2017-04-04. The lookups 
were based on org.Hs.eg.db version 3.4.0.
}
\examples{

data(jetsetscores)
head(jetsetscores)
summary(jetsetscores)

}
\references{
Qiyuan Li, Nicolai J. Birkbak, Balazs Gyorffy, Zoltan Szallasi 
and Aron C. Eklund. (2011) Jetset: selecting the optimal microarray probe 
set to represent a gene. BMC Bioinformatics. 12:474.
}
\author{
Bertrand Frederic
}
\keyword{datasets}
