% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{M}
\alias{M}
\title{Simulated microarray.}
\description{
Simulated M, microarray.
}
\examples{

data(M)
head(M)
str(M)

}
\author{
Bertrand Frederic, Myriam Maumy-Bertrand.
}
\keyword{datasets}
