\name{micropredict-class}
\Rdversion{1.1}
\docType{class}
\alias{micropredict-class}


\title{Class \code{"micropred"}}
\description{
2254
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("micropred", ...)}.
%%  ~~ describe objects here ~~ 
}

\references{
%%  ~~put references to the literature/web site here~~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
}
\examples{
showClass("network")
}
\keyword{classes}
