% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patterns-package.R
\docType{class}
\name{micropredict-class}
\alias{micropredict-class}
\title{Class \code{"micropredict"}}
\description{
The \code{"micropredict"} class
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("micropred", ...)}.
}

\examples{

showClass("micropredict")

}
\author{
Bertrand Frederic, Myriam Maumy-Bertrand.
}
\keyword{classes}
