% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Net}
\alias{Net}
\title{Simulated network for examples.}
\description{
Simulated network.
}
\examples{

data(Net)
str(Net)

}
\author{
Bertrand Frederic, Myriam Maumy-Bertrand.
}
\keyword{datasets}
