% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patterns-package.R
\docType{class}
\name{network-class}
\alias{network-class}
\title{Class \code{"network"}}
\description{
The \code{"network"} class
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("network", ...)}.
}

\examples{

showClass("network")

}
\author{
Bertrand Frederic, Myriam Maumy-Bertrand.
}
\keyword{classes}
