% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{network}
\alias{network}
\title{A example of an inferred network (4 groups case).}
\format{
The format is: chr "network"
}
\description{
This dataset is a network example with 102 nodes, 4 times and 4 groups.
}
\details{
A network class object [package "Patterns"] with 6 slots.
}
\examples{

data(network)
str(network)
plot(network)

}
\author{
Bertrand Frederic, Myriam Maumy-Bertrand.
}
\keyword{datasets}
