% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{network2gp}
\alias{network2gp}
\title{A example of an inferred cascade network (2 groups case).}
\format{
The format is: chr "network2gp"
}
\description{
This dataset is a cascade network example with 53 nodes, 4 times and 2
groups.
}
\details{
A network class object [package "Patterns"] with 6 slots.
}
\examples{

data(network2gp)
str(network2gp)
plot(network2gp)

}
\keyword{datasets}
