% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{networkCascade}
\alias{networkCascade}
\title{A example of an inferred cascade network (4 groups case).}
\format{
The format is: chr "networkCascade"
}
\description{
This dataset is a cascade network example with 102 nodes, 4 times and 4
groups.
}
\details{
A network class object [package "Patterns"] with 6 slots.
}
\examples{

data(networkCascade)
str(networkCascade)
plot(networkCascade)

}
\author{
Bertrand Frederic, Myriam Maumy-Bertrand.
}
\keyword{datasets}
