% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patterns-package.R
\docType{class}
\name{omics_array-class}
\alias{omics_array-class}
\title{Class \code{"omics_array"}}
\description{
The \code{"omics_array"} class
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("omics_array", ...)}.
}

\examples{

showClass("omics_array")

}
\author{
Bertrand Frederic, Myriam Maumy-Bertrand.
}
\keyword{classes}
