% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patterns-package.R
\docType{class}
\name{omics_network-class}
\alias{omics_network-class}
\title{Class \code{"omics_network"}}
\description{
The \code{"omics_network"} class
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("omics_network", ...)}.
}

\examples{

showClass("omics_network")

}
\author{
Bertrand Frederic, Myriam Maumy-Bertrand.
}
\keyword{classes}
