% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patterns-package.R
\docType{class}
\name{omics_predict-class}
\alias{omics_predict-class}
\title{Class \code{"omics_predict"}}
\description{
The \code{"omics_predict"} class
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("omics_predict", ...)}.
}

\examples{

showClass("omics_predict")

}
\author{
Bertrand Frederic, Myriam Maumy-Bertrand.
}
\keyword{classes}
