% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro_array.R
\docType{methods}
\name{print-methods}
\alias{print-methods}
\alias{print,ANY-method}
\alias{print,micro_array-method}
\alias{print,network-method}
\title{\code{Print} methods}
\usage{
\S4method{print}{micro_array}(x, ...)
}
\arguments{
\item{x}{an object of class micro-array or network}

\item{...}{additional parameters}
}
\description{
Methods for generic function \code{print}
}
\section{Methods}{
 \describe{
\item{list("signature(x = \"ANY\")")}{ }
\item{list("signature(x = \"micro_array\")")}{ Print an object of class micro_array }
\item{list("signature(x = \"network\")")}{ Print an object of class network } 
}
}

\author{
Bertrand Frederic, Myriam Maumy-Bertrand.
}
\keyword{methods}
