% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{replaceBand}
\alias{replaceBand}
\title{Replace matrix values by band.}
\usage{
replaceBand(a, b, k)
}
\arguments{
\item{a}{The matrix to be replaced}

\item{b}{The matrix with the replacement values}

\item{k}{The extend of the replacement: 0 (diagonal only), 1 (diagonal and
first extra diagonal), in general an entry is replaced if abs(row(a) -
col(a)) <= k}
}
\value{
A matrix (same size as a)
}
\description{
F matrices utility function.
}
\examples{

a=matrix(1:9,3,3)
b=matrix(0,3,3)
replaceBand(a,b,0)
replaceBand(a,b,1)
replaceBand(a,b,2)

}
\author{
Bertrand Frederic, Myriam Maumy-Bertrand.
}
\keyword{manip}
