% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Selection}
\alias{Selection}
\title{Selection of genes.}
\description{
20 (at most) genes with differential expression at t1, 20 (at most) genes
with differential expression at t2, 20 (at most) genes with differential
expression at t3, 20 (at most) genes with differential expression at t4 et
20 (at most) genes with global differential expression were selected.
}
\examples{

data(Selection)
head(Selection)
summary(Selection,3)

}
\author{
Bertrand Frederic, Myriam Maumy-Bertrand.
}
\keyword{datasets}
