% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_array.R, R/omics_network.R
\docType{methods}
\name{show-methods}
\alias{show-methods}
\alias{show,ANY-method}
\alias{show,omics_array-method}
\alias{show,omics_network-method}
\title{\code{Show} methods}
\usage{
\S4method{show}{omics_array}(object)

\S4method{show}{omics_network}(object)
}
\arguments{
\item{object}{an object of class omics-array or omics_network}
}
\description{
Methods for generic function \code{show}
}
\section{Methods}{
 \describe{
\item{list("signature(object = \"ANY\")")}{ }
\item{list("signature(object = \"omics_array\")")}{ Print an object of class omics_array }
\item{list("signature(object = \"omics_network\")")}{ Print an object of class omics_network } 
}
}

\author{
Bertrand Frederic, Myriam Maumy-Bertrand.
}
\keyword{methods}
