% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{infos}
\alias{infos}
\title{Details on some probesets of the affy_hg_u133_plus_2 platform.}
\format{
The format is: chr "infos"
}
\description{
Dataset with information on the affy_hg_u133_plus_2 platform such as
probeset name (affy_hg_u133_plus_2), ensembl_gene_id, entrezgene,
hgnc_symbol, chromosome_name, start_position, end_position and band.
}
\details{
Data.frame with 8859 rows and 8 variables.
}
\examples{

data(infos)

}
\author{
Bertrand Frederic, Myriam Maumy-Bertrand.
}
\keyword{datasets}
