\name{micropredict-class}
\Rdversion{1.1}
\docType{class}
\alias{micropredict-class}


\title{Class \code{"micropred"}}
\description{
2254
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("micropred", ...)}.
}

\examples{
showClass("network")
}
\keyword{classes}
