% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Net_inf_PL}
\alias{Net_inf_PL}
\title{Reverse-engineered network of the M and Net simulated data.}
\description{
The reverse-engineered network with the `Patterns` package using the
fitfun="LASSO" default function and a cascade network setting.
}
\examples{

data(Net_inf_PL)
str(Net_inf_PL)

}
\author{
Bertrand Frederic, Myriam Maumy-Bertrand.
}
\keyword{datasets}
