\name{network}
\alias{network}
\docType{data}
\title{
A example of an inferred network (4 groups case).
}
\description{
This dataset is a network example with 102 nodes, 4 times and 4 groups.
}
\usage{data("network")}
\format{
  The format is:
 chr "network"
}
\details{
A network class object [package "Patterns"] with 6 slots.
}
\author{
Bertrand Frederic, Myriam Maumy-Bertrand.
}

\examples{
data(network)
str(network)
plot(network)
}
\keyword{datasets}
