\name{networkCascade}
\alias{networkCascade}
\docType{data}
\title{
A example of an inferred cascade network (4 groups case).
}
\description{
This dataset is a cascade network example with 102 nodes, 4 times and 4 groups.
}
\usage{data("networkCascade")}
\format{
  The format is:
 chr "networkCascade"
}
\details{
A network class object [package "Patterns"] with 6 slots.
}
\author{
Bertrand Frederic, Myriam Maumy-Bertrand.
}

\examples{
data(networkCascade)
str(networkCascade)
plot(networkCascade)
}
\keyword{datasets}
