\name{Patterns-package}
\alias{Patterns-package}
\alias{Patterns}
\docType{package}
\title{The Patterns Package}
\description{
The Patterns is a modelisation tool allowing gene selection, reverse engineering, and prediction.
}
\details{
\tabular{ll}{
Package: \tab Cascade\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-03-25\cr
License: \tab GNU 2.0\cr
Depends: \tab methods\cr
}

}
\author{
This package has been written by Nicolas Jung and Frederic Bertrand in collaboration with Myriam Maumy-Bertrand and Laurent Vallat.
Maintainer:  <njung@math.unistra.fr>
}
\references{
Vallat, L., Kemper, C. A., Jung, N., Maumy-Bertrand, M., Bertrand, F., Meyer, N., ... & Bahram, S. (2013). Reverse-engineering the genetic circuitry of a cancer cell with predicted intervention in chronic lymphocytic leukemia. \emph{Proceedings of the National Academy of Sciences}, 110(2), 459-464.

}

\keyword{ package }

