\name{position-methods}
\docType{methods}
\alias{position-methods}
\alias{position,network-method}
\title{Returns the position of edges in the network}
\description{
Returns the position of edges in the network
}
\section{Methods}{
\describe{

\item{\code{signature(net = "network")}}{
Returns a matrix with the position of the node. This matrix can then be used as an argument in the plot function.
}
}}
\keyword{methods}

