% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro_array.R
\docType{methods}
\name{summary-methods}
\alias{summary-methods}
\alias{summary,ANY-method}
\alias{summary,micro_array-method}
\title{\code{Summary} methods}
\usage{
\S4method{summary}{micro_array}(object, nb.graph = NULL, ...)
}
\arguments{
\item{object}{an object of class micro-array}

\item{nb.graph}{(optionnal) choose the graph to plot. Displays all graphs by default.}

\item{...}{additional parameters.}
}
\description{
Methods for function \code{summary}
}
\section{Methods}{
 \describe{

\item{list("signature(object = \"ANY\")")}{ %% ~~describe this method here~~
}

\item{list("signature(object = \"micro_array\")")}{ %% ~~describe this
method here~~ } }
}

\author{
Bertrand Frederic, Myriam Maumy-Bertrand.
}
\keyword{methods}
