\name{unionMicro-methods}
\docType{methods}
\alias{unionMicro-methods}
\alias{unionMicro,micro_array,micro_array-method}
\title{Makes the union between two micro_array objects.}
\description{
Makes the union between two micro_array objects.
}
\section{Methods}{
\describe{

\item{\code{signature(M1 = "micro_array", M2 = "micro_array")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
